requirements.txt
asgiref==3.11.1
Django==6.0.3
sqlparse==0.5.5
tzdata==2025.3
psycopg2-binary>=2.9
python-dotenv>=1.0
openai>=1.0
django-allauth[socialaccount]
